Aula 33  - Dialog de Confirmação

+++ src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/err
 
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -42,16 +43,26 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course) {
-    this.courseService.remove(course.id).subscribe({
-      next: () => {
-        this.refresh();
-        this._snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 3000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',
-        });
-      },
-      error: () => this.onError('Erro ao tentar remover curso!'),
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certexa que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe({
+      next: (result: boolean) => {
+        if (result) {
+          this.courseService.remove(course.id).subscribe({
+            next: () => {
+              this.refresh();
+              this._snackBar.open('Curso removido com sucesso!', 'X', {
+                duration: 3000,
+                verticalPosition: 'top',
+                horizontalPosition: 'center',
+              });
+            },
+            error: () => this.onError('Erro ao tentar remover curso!'),
+          });
+        }
+      }
     });
   }
